Make concat also work on lists of bytes.
Unfortunately this requires manually writing out the mapped function
into the resulting parser, as the `join` method needs to be bound to
the same type as the elements to be joined. Because these elements can
be an empty list, this type needs to be inferred from the type of the
input stream.